# docs/api/openapi/member.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Member Service API",
    "description": "회원 관리 서비스 API 문서",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "API Gateway"
    },
    {
      "url": "http://localhost:8083",
      "description": "로컬 개발 서버 (직접 접근)"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/auth/email-verifications": {
      "post": {
        "tags": [
          "이메일 인증"
        ],
        "summary": "이메일 인증 발송",
        "description": "회원가입용 이메일 인증 코드를 발송합니다.",
        "operationId": "sendEmailVerification",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendEmailVerificationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseEmailVerificationSendResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/email-verifications/confirm": {
      "post": {
        "tags": [
          "이메일 인증"
        ],
        "summary": "이메일 인증 확인",
        "description": "이메일 인증 코드를 검증하고 회원을 활성화합니다.",
        "operationId": "confirmEmailVerification",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmEmailVerificationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseEmailVerificationConfirmResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/email-verifications/auto-login": {
      "post": {
        "tags": [
          "이메일 인증"
        ],
        "summary": "이메일 인증 자동 로그인",
        "description": "이메일 인증 후 발급된 1회용 토큰으로 자동 로그인을 완료합니다.",
        "operationId": "autoLoginAfterEmailVerification",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmailVerificationAutoLoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseEmailVerificationAutoLoginResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
...
      "SendEmailVerificationRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "minLength": 1
          }
        },
        "required": [
          "email"
        ]
      },
      "ApiResponseEmailVerificationSendResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "$ref": "#/components/schemas/EmailVerificationSendResponse"
          },
          "error": {
            "$ref": "#/components/schemas/ErrorResponse"
          }
        }
      },
      "EmailVerificationSendResponse": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "purpose": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ConfirmEmailVerificationRequest": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "minLength": 1
          }
        },
        "required": [
          "token"
        ]
      },
      "ApiResponseEmailVerificationConfirmResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "$ref": "#/components/schemas/EmailVerificationConfirmResponse"
          },
          "error": {
            "$ref": "#/components/schemas/ErrorResponse"
          }
        }
      },
      "EmailVerificationConfirmResponse": {
        "type": "object",
        "properties": {
          "memberId": {
            "type": "string",
            "format": "uuid"
          },
          "email": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING_VERIFICATION",
              "ACTIVE",
              "SUSPENDED",
              "WITHDRAWN",
              "DELETED"
            ]
          },
          "autoLoginToken": {
            "type": "string"
          },
          "autoLoginTokenExpiresInSeconds": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "EmailVerificationAutoLoginRequest": {
        "type": "object",
        "properties": {
          "autoLoginToken": {
            "type": "string",
            "minLength": 1
          }
        },
        "required": [
          "autoLoginToken"
        ]
      },
      "ApiResponseEmailVerificationAutoLoginResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "$ref": "#/components/schemas/EmailVerificationAutoLoginResponse"
          },
          "error": {
            "$ref": "#/components/schemas/ErrorResponse"
          }
        }
      },
      "EmailVerificationAutoLoginResponse": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string"
          },
          "refreshToken": {
            "type": "string"
          },
          "tokenType": {
            "type": "string"
          },
          "accessTokenExpiresIn": {
            "type": "integer",
            "format": "int64"
          },
          "refreshTokenExpiresIn": {
            "type": "integer",
            "format": "int64"
          },
          "sessionId": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}